scraper/requirements.txt
flask==3.0.0
flask-cors==4.0.0
selenium==4.16.0
pymongo==4.6.1
waitress==2.1.2
python-dotenv==1.0.0